Decoupling of code is complete.
